Make EXTERNALIZE_CLASSES mechanism actually work (add the extern part for Point and the cpp part for Halfspace). Remove compilation of number types that cause a bunch of compilation errors. Fix other compilation errors.

+++ src/hypro/datastructures/Halfspace.tpp
@@ -143,7 +143,7 @@ bool Halfspace<Number>::intersection( Number& _result, const Point<Number>& _vec
 }
 
 template <typename Number>
-Halfspace<Number> Halfspace<Number>::projectOn( const std::vector<unsigned>& dimensions ) const {
+Halfspace<Number> Halfspace<Number>::projectOn( const std::vector<size_t>& dimensions ) const {
 	if ( dimensions.empty() ) {
 		return Halfspace<Number>();
 	}
